#637: Split component content into sub-components
